requirements.txt
requests
PyYAML
beautifulsoup4
psycopg2-binary
emoji
pandas
requests
tabulate
tqdm
redis
python-telegram-bot
nest-asyncio